+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixes
 
 - Fix an issue that the file/folder in the file browser was not removed even though when it was deleted in the actual file storage.
+- Fix an issue that the Donation settings pane displayed an unnecessary alert when the user cancels an action.